Start moving content from nav tab to new files
