+++ src/devkit_ui/devkit_ui/constants.py
@@ -1,0 +1,1 @@
+NODE_NAME = 'nicegui'
