+++ src/devkit_ui/devkit_ui/logging.py
@@ -1,0 +1,5 @@
+import rclpy.logging
+
+from devkit_ui.constants import NODE_NAME
+
+app_logger = rclpy.logging.get_logger(NODE_NAME)
