+++ src/devkit_ui/devkit_ui/stores/global_store.py
@@ -1,0 +1,4 @@
+class GlobalStore:
+    soft_estop_active: bool = False
+
+global_store = GlobalStore()
