+++ src/devkit_ui/devkit_ui/stores/run_store.py
@@ -1,0 +1,11 @@
+from nicegui import ui
+
+
+class RunStore:
+    pose_lbl: str = '—'
+
+    map_html: ui.html | None = None
+    robot_html: ui.html | None = None
+
+
+run_store = RunStore()
